Refactor serializers.py
