--- calyx/src/courses/serializers.py
@@ -1,363 +1,0 @@
-import functools
-
-from django.conf import settings
-from django.contrib.auth import get_user_model, password_validation
-from django.core import exceptions
-from django.db import IntegrityError, transaction
-from rest_framework import serializers
-from drf_yasg.utils import swagger_serializer_method
-
-from .models import Course, Year, Config, Lab, Rank, get_config_cache
-from .status import StatusMessage
-
-User = get_user_model()
-
-
-@functools.lru_cache(maxsize=None)
-def get_pin_code_validators():
-    return password_validation.get_password_validators(settings.PIN_CODE_VALIDATORS)
-
-
-class UserSerializer(serializers.ModelSerializer):
-    """
-    ユーザオブジェクトのシリアライザ
-    """
-
-    # NestedViewMixinで使用される親要素のフィルタリング．
-    # dirty hackっぽいが，Course一覧からプライマリキーでフィルタをかけて1つに絞り，参加するユーザ一覧を取得する
-    # `course_pk`がURLパラメータのキー，`pk`が絞り込みに使用するキー
-    parent_lookup_kwargs = {
-        'course_pk': 'pk'
-    }
-
-    class Meta:
-        model = User
-        fields = ("pk", "username", "email", "screen_name", "gpa")
-        extra_kwargs = {
-            "username": {"read_only": True},
-            "email": {"read_only": True},
-            "screen_name": {"read_only": True},
-            "gpa": {"read_only": True}
-        }
-
-    def _rm_from_represent(self, represent: dict, config_cache: dict, cache_key: str, to_del: str) -> dict:
-        """課程の設定のキーと削除対象のキーを受取って存在すれば削除"""
-        if not config_cache.get(cache_key, True):
-            if to_del in represent:
-                represent.pop(to_del)
-        return represent
-
-    def to_representation(self, instance):
-        """課程の設定に応じて返すパラメータを決定する"""
-        represent = super(UserSerializer, self).to_representation(instance)
-        course = self.context.get('course', None)
-        if course:
-            config = get_config_cache(course.pk)
-            represent = self._rm_from_represent(represent, config, 'show_gpa', 'gpa')
-            represent = self._rm_from_represent(represent, config, 'show_username', 'screen_name')
-        return represent
-
-
-class RankListSerializer(serializers.ListSerializer):
-    """希望順位をまとめて作成/更新するシリアライザ"""
-
-    def validate(self, attrs):
-        config = self.context['course'].config  # type: Config
-        if len(attrs) != config.rank_limit:
-            raise serializers.ValidationError(
-                {'non_field_errors': f'希望順位の提出数は{config.rank_limit}個である必要があります．'}
-            )
-        labs = [attr['lab'] for attr in attrs]
-        if len(labs) != len(list(set(labs))):
-            return serializers.ValidationError(
-                {'non_field_errors': '同じ研究室を複数指定することはできません．'}
-            )
-        return super(RankListSerializer, self).validate(attrs)
-
-    def create(self, validated_data):
-        user = self.context['request'].user  # type: User
-        course = self.context['course']  # type: Course
-        rank_list = []
-        with transaction.atomic():
-            for i, data in enumerate(validated_data):
-                rank, _ = Rank.objects.update_or_create(
-                    user=user, course=course, order=i, defaults={'lab': data['lab']}
-                )
-                rank_list.append(rank)
-        return rank_list
-
-
-class RankSerializer(serializers.ModelSerializer):
-    """希望順位を作成するシリアライザ"""
-    lab = serializers.PrimaryKeyRelatedField(
-        queryset=Lab.objects.select_related('course').all(),
-        error_messages={'does_not_exist': "指定された研究室は見つかりませんでした．"},
-    )
-
-    class Meta:
-        model = Rank
-        fields = ('lab',)
-        list_serializer_class = RankListSerializer
-
-
-class RankPerLabListSerializer(serializers.ListSerializer):
-    """研究室ごとの希望順位を表示するシリアライザ"""
-
-    def to_representation(self, data):
-        """
-        [[第1志望のユーザのリスト],[第2志望のユーザのリスト],...]
-        :param data:
-        :return:
-        """
-        course = self.context['course']
-        config = get_config_cache(course.pk)
-        rank_per_order = list()
-        for order in range(config["rank_limit"]):
-            rank_per_order.append(
-                [self.child.to_representation(rank) for rank in data.filter(order=order).all()]
-            )
-        return rank_per_order
-
-
-class RankPerLabSerializer(serializers.ModelSerializer):
-    """
-    希望順位のシリアライザ
-    """
-
-    user = UserSerializer(read_only=True)
-
-    parent_lookup_kwargs = {
-        'lab_pk': 'lab_id'
-    }
-
-    class Meta:
-        model = Rank
-        fields = ('user',)
-        list_serializer_class = RankPerLabListSerializer
-
-    def to_representation(self, instance):
-        """ユーザ情報のみを返却する"""
-        data = super(RankPerLabSerializer, self).to_representation(instance)
-        return data['user']
-
-
-class LabListCreateSerializer(serializers.ListSerializer):
-    """
-    複数の研究室を一括で作成するシリアライザ
-    """
-
-    def to_internal_value(self, data):
-        course = self.context.get("course")
-        for i in range(len(data)):
-            data[i]['course'] = course
-        return data
-
-    def validate(self, data):
-        name_set = [d['name'] for d in data]
-        without_dups = list(set(name_set))
-        if len(name_set) != len(without_dups):
-            raise serializers.ValidationError({'name': '重複した名前は使用できません．'})
-        return data
-
-    def create(self, validated_data):
-        course = self.context.get("course")
-        labs = [Lab(**data) for data in validated_data]
-        Lab.objects.bulk_create(labs)
-        return Lab.objects.filter(course_id=course.pk).all()
-
-
-class LabAbstractSerializer(serializers.ModelSerializer):
-    """
-    希望順位を含まない研究室のシリアライザ
-    """
-
-    course = serializers.PrimaryKeyRelatedField(queryset=Course.objects.select_related('year').all(),
-                                                write_only=True, required=False)
-
-    parent_lookup_kwargs = {
-        'course_pk': 'course_id'
-    }
-
-    class Meta:
-        model = Lab
-        fields = ("pk", "name", "capacity", "course")
-        list_serializer_class = LabListCreateSerializer
-
-
-class LabSerializer(LabAbstractSerializer):
-    """
-    研究室のシリアライザ
-    """
-
-    rank_set = RankPerLabSerializer(many=True, read_only=True)
-
-    class Meta:
-        model = Lab
-        fields = ("pk", "name", "capacity", "course", "rank_set")
-
-    def validate_capacity(self, obj):
-        if obj < 0:
-            raise serializers.ValidationError({"capacity": "許容人数は0人以上である必要があります．"})
-        return obj
-
-
-class ConfigSerializer(serializers.ModelSerializer):
-    """
-    課程ごとの表示設定のシリアライザ．
-    """
-
-    parent_lookup_kwargs = {
-        'course_pk': 'course_id'
-    }
-
-    class Meta:
-        model = Config
-        fields = ("show_gpa", "show_username", "rank_limit")
-
-
-class CourseSerializer(serializers.ModelSerializer):
-    """
-    課程のシリアライザ．年度はリレーション先から年度の数字のみを取得して返す．
-    所属するユーザはそのプライマリキーと学生IDのみを返す．
-    """
-
-    users = serializers.SerializerMethodField(read_only=True)
-    year = serializers.IntegerField(source='year.year')
-    is_admin = serializers.SerializerMethodField(read_only=True)
-    config = ConfigSerializer(required=False)
-
-    class Meta:
-        model = Course
-        fields = ("pk", "name", "users", "pin_code", "year", "is_admin", "config")
-        extra_kwargs = {
-            'pin_code': {'write_only': True}
-        }
-
-    def create(self, validated_data):
-        # yearは{'year': {'year': 20xx}}という辞書になっている
-        validated_data['year'] = validated_data['year']['year']
-        try:
-            course = Course.objects.create_course(**validated_data)
-            if 'request' in self.context.keys():
-                course.join(self.context['request'].user, validated_data['pin_code'])
-            return course
-        except IntegrityError:
-            raise serializers.ValidationError({'name': f'{validated_data["name"]}は既に存在しています．'})
-
-    def update(self, instance, validated_data):
-        try:
-            pin_code = validated_data.pop('pin_code')
-            instance.set_password(pin_code)
-        except KeyError:
-            pass
-        # yearは{'year': {'year': 20xx}}という辞書になっている
-        year = validated_data.get('year', instance.year.year)
-        if isinstance(year, dict):
-            year = year.get('year')
-        validated_data['year'], _ = Year.objects.get_or_create(year=year)
-        # configはOrderedDict
-        config = validated_data.get('config', instance.config)
-        if not isinstance(config, Config):
-            config, _ = Config.objects.update_or_create(course=instance, defaults=config)
-        validated_data['config'] = config
-
-        for key, val in validated_data.items():
-            setattr(instance, key, val)
-        instance.save()
-        return instance
-
-    def validate_pin_code(self, data):
-        try:
-            password_validation.validate_password(data, password_validators=get_pin_code_validators())
-        except exceptions.ValidationError as e:
-            raise serializers.ValidationError(str(e))
-        return data
-
-    @swagger_serializer_method(serializer_or_field=UserSerializer(many=True))
-    def get_users(self, obj):
-        users = obj.users.prefetch_related('groups').all()
-        group_name = obj.admin_group_name
-        return [
-            {
-                'username': user.username,
-                'pk': user.pk,
-                'is_admin': user.groups.filter(name=group_name).exists()
-            } for user in users
-        ]
-
-    def get_is_admin(self, obj) -> bool:
-        try:
-            user = self.context['request'].user
-        except KeyError:
-            return False
-        group_name = obj.admin_group_name
-        return user.groups.filter(name=group_name).exists()
-
-
-class CourseWithoutUserSerializer(serializers.ModelSerializer):
-    """
-    ユーザの情報を含まないCourseSerializer
-    """
-
-    year = serializers.IntegerField(source='year.year')
-
-    class Meta:
-        model = Course
-        fields = ("pk", "name", "year")
-
-    def create(self, validated_data):
-        raise NotImplementedError("This method is not supported in this serializer.")
-
-    def update(self, instance, validated_data):
-        raise NotImplementedError("This method is not supported in this serializer.")
-
-
-class CourseStatusDetailSerializer(serializers.Serializer):
-    """
-    各設定項目を満たしているかどうかを表現するSerializer
-    """
-
-    gpa = serializers.BooleanField(read_only=True)
-    screen_name = serializers.BooleanField(read_only=True)
-    rank_submitted = serializers.BooleanField(read_only=True)
-
-
-class CourseStatusSerializer(serializers.Serializer):
-    """
-    課程の要求する設定項目を満たしているかどうかをチェックするViewのためのSerializer
-    """
-    status = serializers.CharField(read_only=True)
-    status_message = serializers.CharField(read_only=True)
-    detail = CourseStatusDetailSerializer(read_only=True)
-
-    def to_representation(self, instance):
-        """Userインスタンスからステータスをチェックする"""
-        course_pk = self.context['course_pk']
-        status_msg = StatusMessage(instance, course_pk)
-        return super(CourseStatusSerializer, self).to_representation(status_msg)
-
-
-class YearSerializer(serializers.ModelSerializer):
-    """
-    年度の数字と課程の一覧を返す．
-    """
-
-    courses = CourseWithoutUserSerializer(many=True, read_only=True)
-
-    class Meta:
-        model = Year
-        fields = ("pk", "year", "courses")
-
-
-class PINCodeSerializer(serializers.Serializer):
-    """
-    PINコードを取得する
-    """
-
-    pin_code = serializers.CharField(required=True, write_only=True)
-
-    class Meta:
-        fields = ("pin_code",)
-
-    def to_internal_value(self, data):
-        return {'pin_code': data['pin_code']}
